Add movement to the lifts.
When instructed to go to a floor, the lift will now move with
a time delay per floor. This introduces the concept of a
destination for a lift, not just a current floor. Tests have
been updated accordingly.

Destinations aren't yet cleared when the lift gets there, paritally
because int doesn't have a nil value that couldn't otherwise be
meaningful. In the future destinations are likely to be stored in a
list, so this is an accepted drawback for now.