Change a lift's destinations to be a list, rather than a static int.
This replaces GoToFloor with the more representative AddDestination.
Destinations are automatically sorted at time of insertion.

Destinations are now removed when the floor is reached. A still lift
will therefore have an empty destinations list, which also fixes a
hidden bug that creating a lift on a non-zero floor would have a
non-still direction by default.

There's possibly a race condition here, though. Adding a destination
makes use of a 'temp' slice to avoid altering the lift's
destinations slice multiple times and potentially messing with it
while it's in motion. But there may be a problem when a destination
gets added just as a lift leaves that floor, re-adding it to the
destinations list.

There is currently nothing stopping a destination being added that
is in the wrong direction for the lift.