Extend the scheduling algorithm to schedule lifts that are heading in the direction of the call, if a still lift is not available.
This adds a Direction to the CallLift function, matching the idiom
of most lift systems.

This will be less efficient in some cases when there is a moving
lift much closer to the calling floor than the nearest still lift,
but that's an acceptable compromise for now (and better for the
people already in the lift).